completed profile route
